Move filter_candidates.py to filter_orbits.py
